updated module to use django templates
