+++ hs_dataverse/tests/test_dataverse.py
@@ -1,143 +1,13 @@
 import tempfile
 import shutil
-import os
-import json
 
 from django.test import TestCase
 from django.contrib.auth.models import Group
 
-from django_irods.storage import IrodsStorage
-
 from hs_core import hydroshare
-from hs_core.hydroshare import hs_bagit
-from hs_core.hydroshare import get_party_data_from_user
 
-from hs_dataverse.utils import create_metadata_dict
-
-
-# The following two functions are copied from hs_dataverse/management/commands/dataverse.py
-def get_owner_data(resource):
-    """
-    gets the owner metadata for the given resource, and returns it in a dict
-
-    :param resource: the hydroshare resource to get owner_data from
-    :return: a dict containing owner data
-    """
-    owners = list(resource.raccess.owners)
-    if len(owners) == 0:
-        owner_dict = {
-            'username': '',
-            'first_name': '',
-            'last_name': '',
-            'email': '',
-            'organization': ''
-        }
-    else:
-        for o in owners:
-            party = get_party_data_from_user(o)
-            owner_dict = {
-                'username': format(o.username),
-                'first_name': format(o.first_name),
-                'last_name': format(o.last_name),
-                'email': format(o.email),
-                'organization': format(party['organization'])
-            }
-    return owner_dict
-
-
-def get_other_metadata(res, rid):
-    """
-    gets the other metadata for the given resource, and returns it in a dict.
-    other metadata includes extended metadata, funding agency data, contributors, language, doi
-
-    :param resource: the hydroshare resource to get other metadata from
-    :return: a dict containing other metadata
-    """
-    # read extended metadata as key/value pairsi
-    ext_metadata = ''
-    for key, value in list(res.extra_metadata.items()):
-        ext_metadata = ext_metadata + key + ': ' + value + '\n'
-
-    # get funding agency data
-    funding_agency_names = []
-    award_numbers = []
-
-    for a in res.metadata.funding_agencies.all():
-        funding_agency_names.append(a.agency_name)
-        award_numbers.append(a.award_number)
-
-    # get list of contributors
-    contributors = []
-    for c in res.metadata.contributors.all():
-        contributors.append(str(c))
-
-    # if the resource is published, set the doi and update booleans in dict.
-    doi = ''
-    if res.raccess.public:
-        public = True
-    else:
-        public = False
-
-    if res.raccess.published:
-        published = True
-        doi = res.doi
-    else:
-        published = False
-
-    other_metadata_dict = {
-        'rid': rid,
-        'public': public,
-        'published': published,
-        'doi': doi,
-        'extended_metadata_notes': ext_metadata,
-        'language': str(res.metadata.language),
-        'funding_agency_names': funding_agency_names,
-        'award_numbers': award_numbers,
-        'contributors': contributors
-    }
-    return other_metadata_dict
-
-
-def export_bag(self):
-    """
-    exports the bag for the resource, contained in self (self.res)
-
-    :param self: an object with a resource self.res as an attribute
-    :return: a temporary directory containing the temporary files of metadata from the resource's bag
-    """
-    irods_storage_obj = hs_bagit.create_bag_files(self.res)
-    self.assertTrue(isinstance(irods_storage_obj, IrodsStorage))
-    fd = irods_storage_obj.open("{}/data/resourcemetadata.xml".format(self.res.short_id), 'r')
-
-    contents = b""
-    BUFSIZE = 4096
-    block = fd.read(BUFSIZE)
-    while block != b"":
-        # Do stuff with byte.
-        contents += block
-        block = fd.read(BUFSIZE)
-
-    _, mkfile_path = tempfile.mkstemp(prefix='resourcemetadata.xml', dir=self.temp_dir)
-    os.rename(mkfile_path, mkfile_path[:-8])  # remove last 8 characters generated by tempfile
-    mkfile_path = mkfile_path[:-8]
-    with open(mkfile_path, 'wb') as mkfile:
-        mkfile.write(contents)
-
-    owner_dict = get_owner_data(self.res)
-    _, mkfile_path = tempfile.mkstemp(prefix='ownerdata.json', dir=self.temp_dir)
-    os.rename(mkfile_path, mkfile_path[:-8])  # remove last 8 characters generated by tempfile
-    mkfile_path = mkfile_path[:-8]
-    with open(mkfile_path, 'w') as mkfile:
-        json.dump(owner_dict, mkfile)
-
-    other_metadata_dict = get_other_metadata(self.res, self.res.short_id)
-    _, mkfile_path = tempfile.mkstemp(prefix='other_metadata.json', dir=self.temp_dir)
-    os.rename(mkfile_path, mkfile_path[:-8])  # remove last 8 characters generated by tempfile
-    mkfile_path = mkfile_path[:-8]
-    with open(mkfile_path, 'w') as mkfile:
-        json.dump(other_metadata_dict, mkfile)
-
-    return self.temp_dir
+from hs_dataverse.utils import export_bag
+from hs_dataverse.tests.utilities import global_reset
 
 
 class T01CheckMetadata(TestCase):
@@ -149,6 +19,7 @@ class T01CheckMetadata(TestCase):
         :param self: an instance of the T01CheckMetadta object
         :return: nothing
         """
+        global_reset()
         self.group, _ = Group.objects.get_or_create(name='Hydroshare Author')
         self.admin = hydroshare.create_account(
             'admin@gmail.com',
@@ -164,15 +35,6 @@ class T01CheckMetadata(TestCase):
             username='tom',
             first_name='not a dog',
             last_name='last_name_cat',
-            superuser=False,
-            groups=[]
-        )
-
-        self.mouse = hydroshare.create_account(
-            'jerry@gmail.com',
-            username='jerry',
-            first_name='not a cat',
-            last_name='last_name_dog',
             superuser=False,
             groups=[]
         )
@@ -211,15 +73,19 @@ class T01CheckMetadata(TestCase):
             resource_type='CompositeResource',
             owner=self.cat,
             title='all about dog holes',
-            metadata=metadata
+            metadata = metadata
         )
 
         self.res.extra_metadata = {'name': 'John Jackson', 'email': 'jj@gmail.com'}
         self.res.save()
 
-        self.temp_dir = tempfile.mkdtemp(prefix=self.res.short_id, suffix='_dataverse_tempdir', dir='/tmp')
-        self.temp_dir = export_bag(self)
-        self.meta_dict = create_metadata_dict(self.temp_dir)
+        options = {'verbosity': 1, 'settings': None, 'pythonpath': None, 'traceback': False,
+                   'no_color': False, 'resource_ids': [self.res.short_id],
+                   'generate_metadata': False, 'generate_bag': False}
+
+        self.temp_dir = export_bag(self.res.short_id, options)
+        x = evaluate_json_template('hs_dataverse/templates/template.json', self.temp_dir)
+        self.meta_dict = json.loads(x)
 
     def tearDown(self):
         """
@@ -229,8 +95,6 @@ class T01CheckMetadata(TestCase):
         :return: nothing
         """
         super(T01CheckMetadata, self).tearDown()
-        # same pattern as setUp, kill everything I put in
-        # also delete all tempfiles and tempdirs
         shutil.rmtree(self.temp_dir)
 
     def test_title(self):
