+++ hs_dataverse/utils.py
@@ -9,13 +9,23 @@ import sys
 from django.conf import settings
 import googlemaps
 import time
+import requests
+from hs_core.models import BaseResource
+import tempfile
+from django_irods import icommands
+from hs_core.hydroshare import get_party_data_from_user
+from hs_core.hydroshare.hs_bagit import create_bag_files
+from hs_core.tasks import create_bag_by_irods
+from django.template import Context, Template
+from django.utils.safestring import SafeString
+
 
 # global variables
 
 false = False
 true = True
 
-# helper functions
+
 
 
 # if the given string contains a url, returns it. Otherwise, returns the
@@ -51,22 +61,219 @@ def set_field(val):
         return val.text
 
 
-# utility functions
-
-def create_metadata_dict(temp_dir):
+# get owners. An owner is a user. see https://docs.djangoproject.com/en/3.0/ref/contrib/auth/
+def get_owner_data(resource):
     """
-    creates a dict of all the resource's metadata, in the json format specified by dataverse
+    gets the owner metadata for the given resource, and returns it in a dict
+
+    :param resource: the hydroshare resource to get owner_data from
+    :return: a dict containing owner data
+    """
+    owners = list(resource.raccess.owners)
+    if len(owners) == 0:
+        owner_dict = {
+            'username': '',
+            'first_name': '',
+            'last_name': '',
+            'email': '',
+            'organization': ''
+        }
+    else:
+        for o in owners:
+            party = get_party_data_from_user(o)
+
+            owner_dict = {
+                'username': format(o.username),
+                'first_name': format(o.first_name),
+                'last_name': format(o.last_name),
+                'email': format(o.email),
+                'organization': format(party['organization'])
+            }
+    return owner_dict
+
+
+def get_other_metadata(res, rid):
+    """
+    gets the other metadata for the given resource, and returns it in a dict.
+    other metadata includes extended metadata, funding agency data, contributors, language, doi
+
+    :param resource: the hydroshare resource to get other metadata from
+    :return: a dict containing other metadata
+    """
+    # read extended metadata as key/value pairs
+    ext_metadata = ''
+    for key, value in list(res.extra_metadata.items()):
+        ext_metadata = ext_metadata + key + ': ' + value + '\n'
+
+    # get funding agency data
+    funding_agency_names = []
+    award_numbers = []
+
+    for a in res.metadata.funding_agencies.all():
+        funding_agency_names.append(a.agency_name)
+        award_numbers.append(a.award_number)
+
+    # get list of contributors
+    contributors = []
+    for c in res.metadata.contributors.all():
+        contributors.append(str(c))
+
+    # if the resource is published, set the doi and update booleans in dict.
+    doi = ''
+    if res.raccess.public:
+        public = True
+    else:
+        public = False
+
+    if res.raccess.published:
+        published = True
+        doi = res.doi
+    else:
+        published = False
+
+    other_metadata_dict = {
+        'rid': rid,
+        'public': public,
+        'published': published,
+        'doi': doi,
+        'extended_metadata_notes': ext_metadata,
+        'language': str(res.metadata.language),
+        'funding_agency_names': funding_agency_names,
+        'award_numbers': award_numbers,
+        'contributors': contributors
+    }
+
+    return other_metadata_dict
+
+
+def export_bag(rid, options):
+    """
+    exports the bag for the resource with the given resource id, contained in self (self.res)
+    :param rid: the resource id of the resource
+    :param options: any additional command line arguments to be included
+    :return: a temporary directory containing the temporary files of metadata from the resource's bag
+    """
+    requests.packages.urllib3.disable_warnings()
+    try:
+        # database handle
+        resource = BaseResource.objects.get(short_id=rid)
+
+        # instance with proper subclass type and access
+        res = resource.get_content_model()
+        assert res, (resource, resource.content_model)
+        if (res.discovery_content_type != 'Composite'):
+            print("resource type '{}' is not supported. Aborting.".format(res.discovery_content_type))
+            exit(1)
+
+        # create temporary directory
+        mkdir = tempfile.mkdtemp(prefix=rid, suffix='_dataverse_tempdir', dir='/tmp')
+
+        # file handle
+        istorage = res.get_irods_storage()
+        root_exists = istorage.exists(res.root_path)
+        if root_exists:
+            scimeta_path = os.path.join(res.root_path, 'data',
+                                         'resourcemetadata.xml')
+            scimeta_exists = istorage.exists(scimeta_path)
+
+            if scimeta_exists:
+                if icommands.ACTIVE_SESSION:
+                    session = icommands.ACTIVE_SESSION
+                else:
+                    raise KeyError('settings must have irods_global_session set')
+
+                args = ('-')  # redirect to stdout
+                fd = session.run_safe('iget', None, scimeta_path, *args)
+
+                contents = b""
+                BUFSIZE = 4096
+                block = fd.stdout.read(BUFSIZE)
+                while block != b"":
+                    contents += block
+                    block = fd.stdout.read(BUFSIZE)
+
+                _, mkfile_path = tempfile.mkstemp(prefix='resourcemetadata.xml', dir=mkdir)
+                with open(mkfile_path, 'wb') as mkfile:
+                    mkfile.write(contents)
+            else:
+                print("resource metadata {} not found".format(scimeta_path))
+
+            bag_exists = istorage.exists(res.bag_path)
+            dirty = res.getAVU('metadata_dirty')
+            modified = res.getAVU('bag_modified')
+
+            # make sure that the metadata file syncs with the database
+            if dirty or not scimeta_exists or options['generate_metadata']:
+                try:
+                    create_bag_files(res)
+                except ValueError as e:
+                    print(("{}: value error encountered: {}".format(rid, str(e))))
+                    return
+                res.setAVU('metadata_dirty', 'false')
+                res.setAVU('bag_modified', 'true')
+
+            if modified or not bag_exists or options['generate_bag']:
+                create_bag_by_irods(rid)
+                res.setAVU('bag_modified', 'false')
+
+                if icommands.ACTIVE_SESSION:
+                    session = icommands.ACTIVE_SESSION
+                else:
+                    raise KeyError('settings must have irods_global_session set')
+
+                dir = '/'.join([res.root_path, 'data/contents'])
+                istorage = res.get_irods_storage()
+                data = istorage.listdir(dir)
+
+                for file in data[1]:
+                    bag_data_path = '/'.join([res.root_path, 'data/contents', file])
+                    args = ('-')  # redirect to stdout
+                    fd = session.run_safe('iget', None, bag_data_path, *args)
+
+                    contents = b""
+                    block = fd.stdout.read(BUFSIZE)
+                    while block != b"":
+                        # Do stuff with byte.
+                        contents += block
+                        block = fd.stdout.read(BUFSIZE)
+                        _, mkfile_path = tempfile.mkstemp(prefix=file, dir=mkdir)
+                    with open(mkfile_path, 'wb') as mkfile:
+                        mkfile.write(contents)
+
+            owner_dict = get_owner_data(resource)
+            _, mkfile_path = tempfile.mkstemp(prefix='ownerdata.json', dir=mkdir)
+            with open(mkfile_path, 'w') as mkfile:
+                json.dump(owner_dict, mkfile)
+
+            other_metadata_dict = get_other_metadata(res, rid)
+            _, mkfile_path = tempfile.mkstemp(prefix='other_metadata.json', dir=mkdir)
+            with open(mkfile_path, 'w') as mkfile:
+                json.dump(other_metadata_dict, mkfile)
+
+        else:
+            print("Resource with id {} does not exist in iRODS".format(rid))
+    except BaseResource.DoesNotExist:
+        print("Resource with id {} NOT FOUND in Django".format(rid))
+
+    # before returning the temporary directory, rename all the files by
+    # removing the extra characters inserted by mkstemp()
+    for file in os.listdir(mkdir):
+        file_path = '/'.join([mkdir, file])
+        os.rename(file_path, file_path[:-8])  # remove last 8 characters generated by tempfile
+        file_path = file_path[:-8]
+
+    return mkdir
+
+
+def evaluate_json_template(template_path, temp_dir):
+    """
+    gets all of the resource's metadata, then puts it in the template
+    in the json format specified by dataverse
 
     :param temp_dir: the temporary directory containing the resource's bag metadata files
     :return: the metadata dict
     """
     # parse the xml metadata file as an etree
-    with open(os.path.join(sys.path[0], "hs_dataverse",  "template.json"), "r") as read_file:
-        data = json.load(read_file)
-
-    fields = data['datasetVersion']['metadataBlocks']['citation']['fields']
-    geofields = data['datasetVersion']['metadataBlocks']['geospatial']['fields']
-
     meta_path = '/'.join([temp_dir, 'resourcemetadata.xml'])
     tree = ET.parse(meta_path)
     os.remove(meta_path)
@@ -101,299 +308,48 @@ def create_metadata_dict(temp_dir):
 
     relation_tag = '{http://purl.org/dc/elements/1.1/}relation'
 
-    # Define the dicts to be filled in for dataverse fields that accept multiple values
-    other_id_dict = {
-        "otherIdAgency": {
-          "typeName": "otherIdAgency",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": "HydroShare"
-        },
-        "otherIdValue": {
-          "typeName": "otherIdValue",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": ""
-        }
-    }
-
-    author_dict = {
-        "authorName": {
-          "typeName": "authorName",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": "LastAuthor1, FirstAuthor1"
-        },
-        "authorAffiliation": {
-          "typeName": "authorAffiliation",
-          "multiple": false,
-          "typeClass": "primitive",
-        },
-        "authorIdentifierScheme": {
-          "typeName": "authorIdentifierScheme",
-          "multiple": false,
-          "typeClass": "controlledVocabulary",
-          "value": "ORCID"
-        },
-        "authorIdentifier": {
-          "typeName": "authorIdentifier",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": ""
-        }
-    }
-
-    contact_dict = {
-        "datasetContactName": {
-          "typeName": "datasetContactName",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": "LastContact1, FirstContact1"
-        },
-        "datasetContactAffiliation": {
-          "typeName": "datasetContactAffiliation",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": "ContactAffiliation1"
-        },
-        "datasetContactEmail": {
-          "typeName": "datasetContactEmail",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": "ContactEmail1@mailinator.com"
-        }
-    }
-
-    keyword_dict = {
-        "keywordValue": {
-          "typeName": "keywordValue",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": "KeywordTerm1"
-        }
-    }
-
-    related_publications_dict = {
-        "publicationCitation": {
-          "typeName": "publicationCitation",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": ""
-        },
-        "publicationURL": {
-          "typeName": "publicationURL",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": ""
-        }
-    }
-
-    producer_dict = {
-        "producerName": {
-          "typeName": "producerName",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": ""
-        },
-        "producerAffiliation": {
-          "typeName": "producerAffiliation",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": ""
-        },
-        "producerAbbreviation": {
-          "typeName": "producerAbbreviation",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": ""
-        },
-        "producerURL": {
-          "typeName": "producerURL",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": "https://www.hydroshare.org/home/"
-        },
-        "producerLogoURL": {
-          "typeName": "producerLogoURL",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": "https://www.hydroshare.org/static/img/logo-lg.png"
-        }
-    }
-
-    contributor_dict = {
-        "contributorType": {
-          "typeName": "contributorType",
-          "multiple": false,
-          "typeClass": "controlledVocabulary",
-          "value": "Related Person"
-        },
-        "contributorName": {
-          "typeName": "contributorName",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": ""
-        }
-    }
-
-    grant_number_dict = {
-        "grantNumberAgency": {
-          "typeName": "grantNumberAgency",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": ""
-        },
-        "grantNumberValue": {
-          "typeName": "grantNumberValue",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": ""
-        }
-    }
-
-    distributor_dict = {
-        "distributorName": {
-          "typeName": "distributorName",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": "HydroShare"
-        },
-        "distributorAffiliation": {
-          "typeName": "distributorAffiliation",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": "Consortium of Universities for the Advancement of Hydrological Science, Inc (CUAHSI)"
-        },
-        "distributorAbbreviation": {
-          "typeName": "distributorAbbreviation",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": "HydroShare"
-        },
-        "distributorURL": {
-          "typeName": "distributorURL",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": "http://www.hydroshare.org"
-        },
-        "distributorLogoURL": {
-          "typeName": "distributorLogoURL",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": "https://www.hydroshare.org/static/img/logo-lg.png"
-        }
-    }
-
-    software_dict = {
-        "softwareName": {
-          "typeName": "softwareName",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": ""
-        },
-        "softwareVersion": {
-          "typeName": "softwareVersion",
-          "multiple": false,
-          "typeClass": "primitive",
-          "value": ""
-        }
-    }
-
-    geo_coverage_dict = {
-            "country": {
-              "typeName": "country",
-              "multiple": false,
-              "typeClass": "controlledVocabulary",
-              "value": ""
-            },
-            "state": {
-              "typeName": "state",
-              "multiple": false,
-              "typeClass": "primitive",
-              "value": ""
-            },
-            "city": {
-              "typeName": "city",
-              "multiple": false,
-              "typeClass": "primitive",
-              "value": ""
-            },
-            "otherGeographicCoverage": {
-              "typeName": "otherGeographicCoverage",
-              "multiple": false,
-              "typeClass": "primitive",
-              "value": ""
-            }
-    }
-
-    bounding_box_dict = {
-                "westLongitude": {
-                  "typeName": "westLongitude",
-                  "multiple": false,
-                  "typeClass": "primitive",
-                  "value": ''
-                },
-                "eastLongitude": {
-                  "typeName": "eastLongitude",
-                  "multiple": false,
-                  "typeClass": "primitive",
-                  "value": ''
-                },
-                "northLongitude": {
-                  "typeName": "northLongitude",
-                  "multiple": false,
-                  "typeClass": "primitive",
-                  "value": ''
-                },
-                "southLongitude": {
-                  "typeName": "southLongitude",
-                  "multiple": false,
-                  "typeClass": "primitive",
-                  "value": ''
-                }
-    }
-
     # Using the tags, extract fields into local variables
     title = set_field(root.find(".//%s" % title_tag))
     abstract = set_field(root.find(".//%s" % abstract_tag))
 
     alt_url = set_field(root.find(".//%s" % hs_identifier_tag))
 
-    author_vals = []
+    author_dict = {'name': '', 'affiliation': '', 'id_scheme': '', 'identifier': ''} 
+    authors = []
     for creator in root.findall(".//%s" % creator_tag):
-        author_vals.append(copy.deepcopy(author_dict))
+        authors.append(copy.deepcopy(author_dict))
 
-    contact_vals = []
-    for i, author in enumerate(author_vals):
-        author['authorName']['value'] = set_field(root.find(
+    contact_dict = {'name': '', 'affiliation': '', 'email': ''}
+    contacts = []
+    for i, author in enumerate(authors):
+        author['name'] = set_field(root.find(
             ".//%s[%s]/%s/%s" % (creator_tag, i + 1, description_tag, name_tag)))
-        author['authorAffiliation']['value'] = set_field(root.find(
+        author['affiliation'] = set_field(root.find(
             ".//%s[%s]/%s/%s" % (creator_tag, i + 1, description_tag, org_tag)))
-        author['authorIdentifier']['value'] = find_url(set_field(root.find(
+        author['id_scheme'] = 'ORCID'
+        author['identifier'] = find_url(set_field(root.find(
             ".//%s[%s]/%s" % (creator_tag, i + 1, description_tag))), alt_url)
 
         if i == 0:  # use the first author as the contact person for the dataset
             contact = copy.deepcopy(contact_dict)
-            contact['datasetContactName']['value'] = set_field(root.find(
+            contact['name'] = set_field(root.find(
                 ".//%s[%s]/%s/%s" % (creator_tag, i + 1, description_tag, name_tag)))
-            contact['datasetContactAffiliation']['value'] = set_field(root.find(
+            contact['affiliation'] = set_field(root.find(
                 ".//%s[%s]/%s/%s" % (creator_tag, i + 1, description_tag, org_tag)))
-            contact['datasetContactEmail']['value'] = set_field(root.find(
+            contact['email'] = set_field(root.find(
                 ".//%s[%s]/%s/%s" % (creator_tag, i + 1, description_tag, email_tag)))
-            if contact['datasetContactName']['value'] == 'None':
-                contact['datasetContactName']['value'] = 'Hydroshare'
-            if contact['datasetContactEmail']['value'] == 'None':
-                contact['datasetContactEmail']['value'] = 'help@cuahsi.org'
-            contact_vals.append(contact)
+            if contact['name'] == 'None':
+                contact['name'] = 'Hydroshare'
+            if contact['email'] == 'None':
+                contact['email'] = 'help@cuahsi.org'
+            contacts.append(contact)
 
-    keyword_vals = []
-    for keyword in root.findall(".//%s" % keyword_tag):
-        keyword_vals.append(copy.deepcopy(keyword_dict))
+    keywords = []
+    for i, keyword in enumerate(root.findall(".//%s" % keyword_tag)):
+        keyword = set_field(root.find(".//%s[%s]" % (keyword_tag, i + 1)))
+        keywords.append(keyword)
 
-    for i, keyword in enumerate(keyword_vals):
-        keyword['keywordValue']['value'] = set_field(root.find(".//%s[%s]" % (keyword_tag, i + 1)))
-
-    subject = ['Earth and Environmental Sciences']
+    subject = 'Earth and Environmental Sciences'
 
     deposit_date_text = set_field(root.find(".//%s/%s/%s" % (date_tag, created_tag, value_tag)))
     if (deposit_date_text != 'None'):
@@ -420,12 +376,13 @@ def create_metadata_dict(temp_dir):
         start_period_date = ''
         end_period_date = ''
 
-    related_publications_vals = []
+    related_publication_dict = {'citation': '', 'url': ''}
+    related_publications = []
     related_resources = []
     for i, related_publication in enumerate(root.findall(".//%s/%s//*" % (relation_tag, description_tag))):
-        related_publications_vals.append(copy.deepcopy(related_publications_dict))
-        related_publications_vals[i]['publicationCitation']['value'] = set_field(related_publication)
-        related_publications_vals[i]['publicationURL']['value'] = find_url(set_field(related_publication), alt_url)
+        related_publications.append(copy.deepcopy(related_publication_dict))
+        related_publications[i]['citation'] = set_field(related_publication)
+        related_publications[i]['url'] = find_url(set_field(related_publication), alt_url)
 
         related_resources.append(related_publication.text)
 
@@ -437,172 +394,125 @@ def create_metadata_dict(temp_dir):
     maps_api_token = getattr(settings, 'MAPS_KEY', '')
     gmaps = googlemaps.Client(key=maps_api_token)
 
-    geo_units = []
-    bounding_box_vals = []
-    geo_coverage_vals = []
+    geo_info = {'country': '', 'state': '', 'city': '', 'other': ''}
+    geo_units = [] 
+    bounding_box = {'north': '', 'south': '', 'east': '', 'west': ''}
     other_geo_info = ''
-
-    old_geo_coverage_dict = geo_coverage_dict
 
     bounding_box_text = set_field(root.find('.//%s/%s' % (box_tag, value_tag)))
     if (bounding_box_text != 'None'):
-        print('bounding_box_text:', bounding_box_text)
         [northlimit, eastlimit, southlimit, westlimit, units, projection] = bounding_box_text.split(';')
-        northlimit = re.sub('northlimit=', '', northlimit)
-        eastlimit = re.sub('eastlimit=', '', eastlimit)
-        westlimit = re.sub('westlimit=', '', westlimit)
-        southlimit = re.sub('southlimit=', '', southlimit)
+        bounding_box['north'] = re.sub('northlimit=', '', northlimit)
+        bounding_box['east'] = re.sub('eastlimit=', '', eastlimit)
+        bounding_box['west'] = re.sub('westlimit=', '', westlimit)
+        bounding_box['south'] = re.sub('southlimit=', '', southlimit)
         geo_units.append(re.sub('units=', '', units))
-
-        for box in root.findall(".//%s" % box_tag):
-            bounding_box_vals.append(copy.deepcopy(bounding_box_dict))
-
-        for i, box in enumerate(bounding_box_vals):
-            box['westLongitude']['value'] = westlimit
-            box['eastLongitude']['value'] = eastlimit
-            box['northLongitude']['value'] = northlimit
-            box['southLongitude']['value'] = southlimit
 
     spot_text = set_field(root.find('.//%s/%s' % (spot_tag, value_tag)))
     if (spot_text != 'None'):
         [name, east, north, units, projection] = spot_text.split(';', 4)
-        east = re.sub('east=', '', east)
-        north = re.sub('north=', '', north)
-        other_geo_info = re.sub('name=', '', name)
+        # for point coordinates, westlimit == eastlimit, and northlimit == southlimit
+        bounding_box['east'] = re.sub('east=', '', east)
+        bounding_box['west'] = re.sub('east=', '', east)
+        bounding_box['north'] = re.sub('north=', '', north)
+        bounding_box['south'] = re.sub('north=', '', north)
         geo_units.append(re.sub(' units=', '', units))
-        reverse_geo_code_result = gmaps.reverse_geocode((north, east))
+        reverse_geo_code_result = gmaps.reverse_geocode((bounding_box['north'], bounding_box['east']))
         if not reverse_geo_code_result:
             type_dict = {'types': []}
             reverse_geo_code_result.append({'address_components': [type_dict]})
 
-        for box in root.findall(".//%s" % spot_tag):
-            bounding_box_vals.append(copy.deepcopy(bounding_box_dict))
-
-        # for point coordinates, westlimit == eastlimit, and northlimit == southlimit
-        for i, box in enumerate(bounding_box_vals):
-            box['westLongitude']['value'] = east
-            box['eastLongitude']['value'] = east
-            box['northLongitude']['value'] = north
-            box['southLongitude']['value'] = north
-
         for comp in reverse_geo_code_result[0]['address_components']:
             if 'country' in comp['types']:
-                geo_coverage_dict['country']['value'] = comp['long_name']
+                geo_info['country'] = comp['long_name']
             if 'administrative_area_level_1' in comp['types']:
-                geo_coverage_dict['state']['value'] = comp['long_name']
+                geo_info['state'] = comp['long_name']
             if 'locality' in comp['types']:
-                geo_coverage_dict['city']['value'] = comp['long_name']
-    geo_coverage_dict['otherGeographicCoverage']['value'] = other_geo_info
-
-    if (geo_coverage_dict != old_geo_coverage_dict):
-        geo_coverage_vals.append(geo_coverage_dict)
+                geo_info['city'] = comp['long_name']
+        geo_info['other'] = re.sub('name=', '', name)
 
     # Extract more fields using the extended metadata from other_metadata.json
     e = dict()
     with open('/'.join([temp_dir, 'other_metadata.json'])) as f:
         e = json.load(f)
 
-    other_id_dict['otherIdValue']['value'] = str(e['rid'])
-
-    notes_text = e['extended_metadata_notes']
-
-    grant_vals = []
+    notes = e['extended_metadata_notes']
+    rid = e['rid']
+    
+    grant_dict = {'name': '', 'number': ''}
+    grant_info = []
     for i, grant in enumerate(e['award_numbers']):
-        grant_info = copy.deepcopy(grant_number_dict)
-        grant_info['grantNumberAgency']['value'] = e['funding_agency_names'][i]
-        grant_info['grantNumberValue']['value'] = e['award_numbers'][i]
-        grant_vals.append(grant_info)
+        g = copy.deepcopy(grant_dict)
+        g['name'] = e['funding_agency_names'][i]
+        g['number'] = e['award_numbers'][i]
+        grant_info.append(g)
 
-    contributor_vals = []
+    contributors = []
     for contributor in e['contributors']:
-        c_dict = copy.deepcopy(contributor_dict)
-        c_dict['contributorName']['value'] = contributor
-        contributor_vals.append(c_dict)
+        contributors.append(contributor)
 
     # Extract more fields using the owner data from ownerdata.json
     o = dict()
     with open('/'.join([temp_dir, 'ownerdata.json'])) as f:
         o = json.load(f)
     contact = copy.deepcopy(contact_dict)
+    producer_dict = {'name': '', 'abreviation': '', 'affiliation': ''}
     if 'username' in o:
-        producer_dict['producerAbbreviation']['value'] = o['username']
+        producer_dict['abbreviation'] = o['username']
     else:
-        producer_dict['producerAbbreviation']['value'] = ''
+        producer_dict['abbreviation'] = ''
 
     if 'first_name' in o and 'last_name' in o:
-        producer_dict['producerName']['value'] = o['last_name'] + ', ' + o['first_name']
-        contact['datasetContactName']['value'] = o['last_name'] + ', ' + o['first_name']
+        producer_dict['name'] = o['last_name'] + ', ' + o['first_name']
+        contact['name'] = o['last_name'] + ', ' + o['first_name']
         depositor = o['last_name'] + ', ' + o['first_name']
     else:
-        producer_dict['producerName']['value'] = ''
-        contact['datasetContactName']['value'] = ''
+        producer_dict['name'] = ''
+        contact['name'] = ''
         depositor = ''
 
     if 'organization' in o:
-        producer_dict['producerAffiliation']['value'] = o['organization']
-        contact['datasetContactAffiliation']['value'] = o['organization']
+        producer_dict['affiliation'] = o['organization']
+        contact['affiliation'] = o['organization']
     else:
-        producer_dict['producerAffiliation']['value'] = ''
-        contact['datasetContactAffiliation']['value'] = ''
+        producer_dict['affiliation'] = ''
+        contact['affiliation'] = ''
 
     if 'email' in o:
-        contact['datasetContactEmail']['value'] = o['email']
+        contact['email'] = o['email']
     else:
-        contact['datasetContactEmail']['value'] = ''
+        contact['email'] = ''
     if contact != contact_dict:
-        contact_vals.append(contact)
+        contacts.append(contact)
 
-    # update the json dict with the field values
-    fields[0]['value'] = title
+    with open(template_path, 'r') as f:
+        template = Template(f.read())
+    context = Context({'title': title,
+                       'alt_url': alt_url,
+                       'rid': rid,
+                       'authors': authors,
+                       'contacts': contacts,
+                       'abstract': abstract,
+                       'modified_date' : last_modified_date,
+                       'subject': subject,
+                       'keywords': keywords,
+                       'related_publications': related_publications,
+                       'notes': notes,
+                       'producers': [producer_dict], 
+                       'contributors': contributors,
+                       'grant_info': grant_info,
+                       'distribution_date': str(datetime.date(datetime.now())),
+                       'depositor': depositor,
+                       'deposit_date': deposit_date,
+                       'start_period_date': start_period_date,
+                       'end_period_date': end_period_date,
+                       'related_materials': related_materials,
+                       'geo_info': geo_info,
+                       'geo_units': geo_units,
+                       'bounding_box': bounding_box
+                      })
 
-    fields[1]['value'] = alt_url
-
-    fields[2]['value'] = [other_id_dict]
-
-    fields[3]['value'] = author_vals
-    fields[4]['value'] = contact_vals
-
-    fields[5]['value'][0]['dsDescriptionValue']['value'] = abstract
-    fields[5]['value'][0]['dsDescriptionDate']['value'] = last_modified_date
-
-    fields[6]['value'] = subject
-
-    fields[7]['value'] = keyword_vals
-
-    fields[8]['value'] = related_publications_vals
-
-    fields[9]['value'] = notes_text
-
-    fields[10]['value'] = [producer_dict]
-
-    fields[11]['value'] = contributor_vals
-
-    fields[12]['value'] = grant_vals
-
-    fields[13]['value'] = [distributor_dict]
-
-    fields[14]['value'] = str(datetime.date(datetime.now()))  # distribution date
-
-    fields[15]['value'] = depositor
-
-    fields[16]['value'] = deposit_date
-
-    fields[17]['value'][0]['timePeriodCoveredStart']['value'] = start_period_date
-    fields[17]['value'][0]['timePeriodCoveredEnd']['value'] = end_period_date
-
-    fields[18]['value'] = ['Composite Resource']
-
-    fields[19]['value'] = [software_dict]
-
-    fields[20]['value'] = related_materials  # related Material
-
-    fields[21]['value'] = []  # related Datasets
-
-    geofields[0]['value'] = geo_coverage_vals
-    geofields[1]['value'] = geo_units
-    geofields[2]['value'] = bounding_box_vals
-
-    return data
+    return template.render(context)
 
 
 def upload_dataset(base_url, api_token, dv, temp_dir):
@@ -615,29 +525,28 @@ def upload_dataset(base_url, api_token, dv, temp_dir):
     :param temp_dir: the temporary directory containing the resource's bag metadata files
     :return: nothing
     """
-    metadata = create_metadata_dict(temp_dir)
+    
+    x = evaluate_json_template('hs_dataverse/templates/template.json', temp_dir)
+    metadata = json.loads(x)
     api = Api(base_url, api_token)
 
-    query_str = '/dataverses/' + dv + '/contents'
-    params = {}
-    resp = api.get_request(query_str, params=params, auth=True)
-
+    # query_str = '/dataverses/' + dv + '/contents'
+    # params = {}
+    # resp = api.get_request(query_str, params=params, auth=True)
+    #   
     # extracting response data in json format
-    dv_data = resp.json()
-
-    num_dv = len(dv_data[u'data'])
-    print(num_dv)  # eventually, there should be a button which allows users to choose locations from among these
+    # dv_data = resp.json()
+    # 
+    # num_dv = len(dv_data[u'data'])
+    # eventually, there should be a button which allows users to choose locations from among these
     # print all dataverse and dataset titles
-    for i in range(0, num_dv):
-        if((dv_data[u'data'][i][u'type'] == 'dataverse')):
-            print("Dataverse: " + dv_data[u'data'][i][u'title'])
-        if((dv_data[u'data'][i][u'type'] == 'dataset')):
-            print("Dataset: " + dv_data[u'data'][i][u'identifier'])
+    # for i in range(0, num_dv):
+    #    if((dv_data[u'data'][i][u'type'] == 'dataverse')):
+    #        print("Dataverse: " + dv_data[u'data'][i][u'title'])
+    #    if((dv_data[u'data'][i][u'type'] == 'dataset')):
+    #        print("Dataset: " + dv_data[u'data'][i][u'identifier'])
 
     r = api.create_dataset(dv, json.dumps(metadata))
-    print(r)
-    print(r.text)
-
     r_dict = json.loads(r.text)
     persistent_id = r_dict['data']['persistentId']
 
@@ -656,6 +565,7 @@ def upload_dataset(base_url, api_token, dv, temp_dir):
         while not done:
             try:
                 api.upload_file(persistent_id, file_path)
+                os.remove(file_path)
                 done = True
             except:
                 count += 1
@@ -664,9 +574,3 @@ def upload_dataset(base_url, api_token, dv, temp_dir):
                     print('Could not upload file: {}'.format(file))
                 else:
                     time.sleep(7)
-
-    # now delete the dataset, as to not fill up the datverse while testing
-    # time.sleep(3)
-    # r2 = api.delete_dataset(persistent_id, is_pid=True, auth=True)
-    # print(r2)
-    # print(r2.text)
